Move out zoomLevels into typed object
