+++ frontend-service/src/components/MapBoxMap.types.tsx
@@ -1,0 +1,8 @@
+import type { Round } from "../Types";
+
+interface MapboxMapProps {
+	roundDetails: Round;
+	handleGuess: (distance: number) => void;
+}
+
+export type { MapboxMapProps };
