+++ frontend-service/src/objects/zoomLevels.type.ts
@@ -1,0 +1,7 @@
+type ZoomLevel = {
+    maxDistance: number;
+    zoomLevel: number;
+    speed: number;
+}
+
+export type { ZoomLevel }
